requirements.txt
pytest>=8.0,<9.0
pypdf>=4.2,<5.0
pyyaml>=6.0,<7.0
presidio-analyzer>=2.2,<3.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
pymysql>=1.1,<2.0
